fix: userID 타입 수정

+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/request/v3/ChatReportRequestDto.java
@@ -18,5 +18,5 @@ public class ChatReportRequestDto {
     private String messageContent;
 
     @NotNull
-    private int reportedUserId;
+    private long reportedUserId;
 }
